+++ payments/RELEASES.md
@@ -98,3 +98,13 @@ Based on:
 - [typescript v2.3.0] payments
 ### Releases
 - [NPM v2.3.0] https://www.npmjs.com/package/@wingspan/payments/v/2.3.0 - payments
+
+## 2024-08-19 01:31:30
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0-oas3.1 ./openapi/payment_speakeasy_edit.yaml
+- Speakeasy CLI 1.374.2 (2.401.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.3.1] payments
+### Releases
+- [NPM v2.3.1] https://www.npmjs.com/package/@wingspan/payments/v/2.3.1 - payments